# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the login endpoint",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Display a listing of the Subscriptions",
                "operationId": "d0f1b367da9e9395814bc66331ce88a1",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of subscriptions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Subscription"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Store a newly created Subscription in storage",
                "operationId": "286ca8bfa118011978c613004009c3f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSubscriptionAPIRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Subscription": {
                "description": "Subscription model",
                "required": [
                    "id",
                    "name",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the subscription",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the subscription",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status of the subscription",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CreateSubscriptionAPIRequest": {
                "description": "Create subscription request",
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "user_id": {
                        "description": "user_id of the subscription",
                        "type": "string"
                    },
                    "plan_id": {
                        "description": "plan_id of the subscription",
                        "type": "string"
                    },
                    "start_date": {
                        "description": "start_date of the subscription",
                        "type": "string",
                        "format": "date"
                    },
                    "end_date": {
                        "description": "end_date of the subscription",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "description": "Status of the subscription",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}